Add Dockerfiles, Makefile for testing against isolated environment
This commit adds a Dockerfile for PHP 5.6.x and PHP 7.1.x versions, allowing for tests, package
builds etc. to run against an isolated, controlled environment. PHP is currently built as a shared
library, in the future we may build as a static library to allow bundling with Go applications.

Also included is a Makefile, containing targets for common operations, such as testing, building
packages etc. Two undocumented targets related to Docker are included, `docker-image` and
`docker-test`. Moving these as documented targets is a future task.