Unify Dockerfiles for easier testing across PHP versions
All PHP builds in Docker are now made against a single-unified Dockerfile, which accepts the PHP
version to target in the `PHP_VERSION` argument. Source files are currently not verified (other than
being downloaded from a secure location, as we will eventually move to using the official `php`
Docker images as a base, as soon as these catch up to latest versions of Debian.