Fix static compilation for Docker environment
It's unfortunate that Debian packages do not contain symbolic links for base libraries to their
versioned counterparts, and we need to install the `-dev` packages for that. All required packages
are now included in the Docker image's runtime dependencies, and the `static` build tag is made to
include the correct ones.