Fix error logging with default engine configuration
By default, configuration values for `log_errors` and `display_errors` depend on how PHP has been
configured, and need to be made consistent in order for functionality to conform to expectations.

Additionally, the list of errors handled has been expanded to `E_ALL` instead of whatever the
default value is.